requirements.txt
numpy>=1.20
scipy>=1.7 
colour-science>=0.4 